Moved collection.common to impl
Moved the collection.common to collection.impl.common since the content is not
public API
